+++ server/src-lib/Hasura/GraphQL/Schema/Relay.hs
@@ -23,6 +23,7 @@ import Hasura.Base.ToErrorValue
 import Hasura.GraphQL.Schema.Backend
 import Hasura.GraphQL.Schema.Common
 import Hasura.GraphQL.Schema.Instances ()
+import Hasura.GraphQL.Schema.NamingCase (NamingCase)
 import Hasura.GraphQL.Schema.Node
 import Hasura.GraphQL.Schema.Options qualified as Options
 import Hasura.GraphQL.Schema.Parser (Kind (..), Parser, memoizeOn)
@@ -110,6 +111,7 @@ nodeField sourceCache = do
   let idDescription = G.Description "A globally unique id"
       idArgument = P.field Name._id (Just idDescription) P.identifier
   stringifyNumbers <- retrieve Options.soStringifyNumbers
+  tCase <- asks getter
   nodeObject <-
     retrieve scSchemaKind >>= \case
       HasuraSchema -> throw500 "internal error: the node field should only be built for the Relay schema"
@@ -130,7 +132,7 @@ nodeField sourceCache = do
           let matchingTables = flip mapMaybe (Map.keys sourceCache) \sourceName ->
                 (sourceName,) <$> findNode @('Postgres 'Vanilla) sourceName tableName parseds
           case matchingTables of
-            [(sourceName, nodeValue)] -> createRootField stringifyNumbers sourceName tableName nodeValue pKeys
+            [(sourceName, nodeValue)] -> createRootField stringifyNumbers sourceName tableName nodeValue pKeys (Just tCase)
             [] -> throwInvalidNodeId $ "no such table found: " <> toErrorValue tableName
             l ->
               throwInvalidNodeId $
@@ -146,7 +148,7 @@ nodeField sourceCache = do
             nodeValue <-
               findNode @b sourceName tableName parseds
                 `onNothing` throwInvalidNodeId ("no table " <> toErrorValue tableName <> " found in source " <> toErrorValue sourceName)
-            createRootField stringifyNumbers sourceName tableName nodeValue pKeys
+            createRootField stringifyNumbers sourceName tableName nodeValue pKeys (Just tCase)
   where
     throwInvalidNodeId :: ErrorMessage -> n a
     throwInvalidNodeId t = P.withKey (J.Key "args") $ P.withKey (J.Key "id") $ P.parseError $ "invalid node id: " <> t
@@ -165,8 +167,9 @@ nodeField sourceCache = do
       TableName b ->
       NodeInfo b ->
       NESeq.NESeq J.Value ->
+      Maybe NamingCase ->
       n (IR.QueryRootField IR.UnpreparedValue)
-    createRootField stringifyNumbers sourceName tableName (NodeInfo sourceConfig perms pKeys fields) columnValues = do
+    createRootField stringifyNumbers sourceName tableName (NodeInfo sourceConfig perms pKeys fields) columnValues tCase = do
       whereExp <- buildNodeIdBoolExp columnValues pKeys
       pure $
         IR.RFDB sourceName $
@@ -186,7 +189,8 @@ nodeField sourceCache = do
                             IR._saOffset = Nothing,
                             IR._saDistinct = Nothing
                           },
-                      IR._asnStrfyNum = stringifyNumbers
+                      IR._asnStrfyNum = stringifyNumbers,
+                      IR._asnNamingConvention = tCase
                     }
 
     -- Craft the 'where' condition of the query by making an `AEQ` entry for